requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
asttokens==3.0.1
attrs==25.4.0
audioread==3.1.0
bokeh==3.8.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
executing==2.2.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_resources==6.5.2
intervaltree==3.2.1
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
miditoolkit==1.0.1
mido==1.3.3
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.15.0
networkx==3.6.1
note-seq==0.0.5
numba==0.63.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
pretty_midi==0.2.11
prompt_toolkit==3.0.52
protobuf==6.33.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
sentry-sdk==2.49.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wandb==0.23.1
wcwidth==0.2.14
xyzservices==2025.11.0
